(Maint) Remove unneeded tests for operatingsystem release patch versions
This commit removes several tests from the operatingsystem structured fact.
Prior to the release of this fact, we had considered adding a "patch" key in the
release section of the fact, but that idea was scrapped in favor of only
reporting major and minor versions.

A reference to this key was accidentally left in the stubbing of the structured
operatingsystem fact logic. It has now been removed to reflect reality.