(FACT-728) Update Windows OS release resolution for > 6.2.x kernels
Currently, Facter only knows how to deal with Windows kernels
up to 6.2.x for parsing out the operating system release value.
Windows 8.1 and 10 use kernel versions 6.3.9600 and 6.4.x
respectively, which cause Facter to fall back to reporting the
kernel version itself as the operating system release, instead
of the strings "8" or "10" as it should.

This commit adds a check for kernel versions starting with "6.3"
or "6.4" to resolve this issue.