(FACT-964) Search archlibdir and sitearchlibdir for libruby
Prior to this commit, Facter only searched `RbConfig['libdir']`
for libruby. However, the ruby2.1 package places libruby
under `RbConfig['archlibdir']` or `sitearchlibdir` on some platforms.
Thus, Facter wasn't able to find libruby on said systems. This commit
adds `archlibdir` and `sitearchlibdir` to the list of paths to check
for libruby.