(FACT-911) Properly detect SPARC systems when resolving dmi fact
Previously, our logic for determining if a system was SPARC or not
was to check if the architecture fact resolved to "sparc." This
is incorrect, as we've generally always set architecture to the
specific CPU type (i.e, "sun4v"). This commit updates this logic
to instead check the hardwareisa fact, which will consistently
resolve as sparc.