(FACT-1416) Use the -c flag when calling zoneadm for Solaris zone facts
Previously, Facter called zoneadm using only the -p flag when
querying the system for zones, which excluded non-running zones.
This is a breaking change from Facter 2, where Facter reported
all zones.

This commit updates the zone resolver to use the -c flag when
calling zoneadm, as we did in Facter 2. In addition, the zone
regex has been modified to match a '-' at the beginning of a
zone entry, as is the output for non-running zones.